renamed SessionLockRepository to AuthLockRepository
